Remove old testing models
Migrate surviving code from app/cdash/app/Models to app/Model.

Update remove_builds test to also check the testoutput table.
