+++ app/Models/Test.php
@@ -10,4 +10,11 @@ class Test extends Model
     protected $fillable = ['name', 'projectid'];
 
     public $timestamps = false;
+
+    const FAILED = 'failed';
+    const PASSED = 'passed';
+    const OTHER_FAULT = 'OTHER_FAULT';
+    const TIMEOUT = 'Timeout';
+    const NOTRUN = 'notrun';
+    const DISABLED = 'Disabled';
 }
